Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/constructors/Worker/AbstractWorker.onerror-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token ')'

PASS AbstractWorker.onerror 
